obj_list fixed, all leaks are freed
