+++ miniRT/Makefile
@@ -10,8 +10,10 @@ SRC_FILES	= main.c \
 			parsing/parsing.c \
 			parsing/parsing_error.c \
 			parsing/fill_objs.c \
+			parsing/fill_scene.c \
 			parsing/check.c \
 			parsing/obj_list.c \
+			parsing/fill_utils.c \
 
 UNAME		=	$(shell uname)
