+++ miniRT/incs/miniRT.h
@@ -20,9 +20,10 @@
 t_mlx	*setup_mlx(void);
 void	events_loop(t_mlx *mlx);
 void	clean_mlx(t_mlx *mlx);
-int		parsing(char **argv, int argc);
+int		parsing(char **argv, int argc, t_parsing *parsing_var);
 int		is_valid_type(char *type);
 void	ft_free(void *to_free);
+void	free_list(t_obj *obj);
 int		check(void *obj, int type);
 
 int		fill_rgb(char *rgb_values, t_rgb *rgb);
@@ -39,6 +40,10 @@ void	fill_by_type(int type, t_parsing *var, char *line);
  */
 void	fill_structure(t_parsing *parsing_var);
 t_obj	*new_obj(int type, double diameter, double height);
+void	fill_scene(int type, t_parsing *var, char *line);
+void	fill_obj(int type, t_parsing *var, char *line);
 void	obj_add_back(t_obj **obj_list, t_obj *to_add);
+void	print_list(t_obj *obj);
+_Bool	is_valid_extension(char *arg);
 
 #endif
