+++ miniRT/libft/Makefile
@@ -24,6 +24,7 @@ SRCS		= ft_memset.c\
 			  ft_calloc.c\
 			  ft_strdup.c\
 			  ft_substr.c\
+			  ft_strcmp.c\
 			  ft_strjoin.c\
 			  ft_strtrim.c\
 			  ft_split.c\
