+++ miniRT/libft/ft_strcmp.c
@@ -1,0 +1,16 @@
+# define EQUAL 1
+# define DIFFERENT 0
+
+_Bool	ft_strcmp(char *s1, char *s2)
+{
+	int i;
+
+	i = 0;
+	while (s1 && s1[i] && s2 && s2[i] && s1[i] == s2[i])
+	{
+		if (s1[i + 1] == 0 && s2[i + 1] == 0)
+			return (EQUAL);
+		i++;
+	}
+	return (DIFFERENT);
+}
