+++ miniRT/libft/libft.h
@@ -66,5 +66,6 @@ char		*ft_strtrim(const char *s1, const char *set);
 char		*ft_substr(const char *s, unsigned int start, size_t len);
 int			ft_tolower(int c);
 int			ft_toupper(int c);
+_Bool		ft_strcmp(char *s1, char *s2);
 
 #endif
